+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 ## [0.0.2]
 
 * Fix didUpdateWidget
+
+## [0.1.0]
+
+* StateLifecycleSubject -> LifecycleSubject
+* StateLifecycleSubject() -> LifecycleSubject.create()